/vocab/en-us: improve user interaction
Previous to this patch, the interaction between the user and mycroft was very rigid when
requesting details about their timetable. Often only one string was accepted and any
adaptation of that string was seen to be garbage. This patch fixes this, somewhat, as it allows
for more variation but can still be improved.

Signed-off-by: Brian <BrianArchbold96@yahoo.ie>